make circle.yml Fixed #297  rename blogbot.js to blogBot.js
