+++ test/test_blogBot.js
@@ -4,7 +4,7 @@
 'use strict';
 
 var assert  = require('assert');
-var bB = require('../controllers/blogbot');
+var bB = require('../controllers/blogBot');
 var tD = require('./test_data');
 if (!global.log) {
     global.log = require('winston');
